Move markup helpers into a shared _markup_utils module
The markup-stripping primitives are pure string functions with no dependency
on TextSegmentMap, and normalize() carried its own copy of the tag-stripping
regex. Both notions of markup -- fragment-tolerant and complete-text -- now
live in one module that _alnum_utils and text_segment_map share.
