+++ src/pipecat/utils/context/word_completion_tracker.py
@@ -8,7 +8,8 @@
 
 from loguru import logger
 
-from pipecat.utils.context.text_segment_map import TextSegmentMap, strip_complete_markup
+from pipecat.utils.context.text_segment_map import TextSegmentMap
+from pipecat.utils.text.transforms._markup_utils import strip_complete_markup
 
 
 class WordCompletionTracker:
